refactor: move history links to own component
