+++ src/components/Products/Products.tsx
@@ -30,7 +30,6 @@ const Products = () => {
     dispatch(updateMinMaxStock());
   }, [dispatch, products]);
 
-  // Чтоб при старте проверялись url параметры и они обновили фильтры, чтоб они прмиенились.
   // Этот УЖАС чтоб на основе фильтров выставлять актуальный url параметры..........
   useEffect(() => {
     const {
